Creating new files for code quality and license
