+++ docs/PARALLEL_IMPLEMENTATION.md
