+++ src/scorer.py
@@ -18,92 +18,12 @@ from .dataset_quality import calculate_dataset_quality_with_timing
 from .ramp_up_time import calculate_ramp_up_time_with_timing
 from .performance_claims import calculate_performance_claims_with_timing
 from .net_score import calculate_net_score_with_timing
+from .license import calculate_license_score_with_timing, license_score_map
+from .code_quality import calculate_code_quality_with_timing
 import subprocess
 import json
 from concurrent.futures import ThreadPoolExecutor, as_completed
 
-# A map of Hugging Face's valid license keywords (pulled from here: https://huggingface.co/docs/hub/en/repositories-licenses)
-# to scores (scores generated by llama4 along with justifications). I have made some tweaks where I disagreed with the AI output.
-license_score_map : dict[str, float] = {
-    "apache-2.0": 0.95,  # Very permissive with minimal restrictions and a patent grant.
-    "mit": 0.98,  # Extremely permissive with almost no restrictions.
-    "bigscience-openrail-m": 0.8,  # Some restrictions on use
-    "creativeml-openrail-m": 0.8,  # Some restrictions on use
-    "bigscience-bloom-rail-1.0": 0.7,  # Some restrictions on use
-    "bigcode-openrail-m": 0.8,  # Some restrictions on use
-    "afl-3.0": 0.85,  # Permissive but with some restrictions on derivative works.
-    "artistic-2.0": 0.8,  # Permissive but with some restrictions on derivative works.
-    "bsl-1.0": 0.85,  # Permissive but with some restrictions on derivative works.
-    "bsd": 0.85,  # Similar to BSD-3-Clause but without patent protection.
-    "bsd-2-clause": 0.9,  # Very permissive with minimal restrictions.
-    "bsd-3-clause": 0.85,  # Permissive but with some restrictions on patent protection.
-    "bsd-3-clause-clear": 0.85,  # Similar to BSD-3-Clause but with clearer language.
-    "c-uda": 0.6,  # Restricted by NVIDIA's proprietary software requirements.
-    "cc0-1.0": 0.99,  # Public domain dedication with no restrictions.
-    "cc-by-2.0": 0.7,  # Permissive but requires attribution.
-    "cc-by-2.5": 0.7,  # Permissive but requires attribution.
-    "cc-by-3.0": 0.7,  # Permissive but requires attribution.
-    "cc-by-4.0": 0.7,  # Permissive but requires attribution.
-    "cc-by-sa-3.0": 0.5,  # Permissive but requires attribution and same-license derivative works.
-    "cc-by-sa-4.0": 0.5,  # Permissive but requires attribution and same-license derivative works.
-    "cc-by-nc-2.0": 0.4,  # Restricted to non-commercial use.
-    "cc-by-nc-3.0": 0.4,  # Restricted to non-commercial use.
-    "cc-by-nc-4.0": 0.4,  # Restricted to non-commercial use.
-    "cc-by-nd-4.0": 0.3,  # Restricted to no derivative works.
-    "cc-by-nc-nd-3.0": 0.2,  # Restricted to non-commercial use and no derivative works.
-    "cc-by-nc-nd-4.0": 0.2,  # Restricted to non-commercial use and no derivative works.
-    "cc-by-nc-sa-2.0": 0.3,  # Restricted to non-commercial use and same-license derivative works.
-    "cc-by-nc-sa-3.0": 0.3,  # Restricted to non-commercial use and same-license derivative works.
-    "cc-by-nc-sa-4.0": 0.3,  # Restricted to non-commercial use and same-license derivative works.
-    "cdla-sharing-1.0": 0.7,  # Permissive but with some restrictions on data sharing.
-    "cdla-permissive-1.0": 0.85,  # Permissive with minimal restrictions.
-    "cdla-permissive-2.0": 0.85,  # Permissive with minimal restrictions.
-    "wtfpl": 1.0,  # Extremely permissive with almost no restrictions.
-    "ecl-2.0": 0.8,  # Permissive but with some restrictions on derivative works.
-    "epl-1.0": 0.7,  # Restricted by Eclipse Foundation's requirements.
-    "epl-2.0": 0.7,  # Restricted by Eclipse Foundation's requirements.
-    "etalab-2.0": 0.7,  # Permissive but with some restrictions on data sharing.
-    "eupl-1.1": 0.6,  # Restricted by European Union's requirements.
-    "eupl-1.2": 0.6,  # Restricted by European Union's requirements.
-    "agpl-3.0": 0.7,  # Restricted to copyleft requirements.
-    "gfdl": 0.6,  # Restricted to copyleft requirements and invariant sections.
-    "gpl": 0.7,  # Restricted to copyleft requirements.
-    "gpl-2.0": 0.7,  # Restricted to copyleft requirements.
-    "gpl-3.0": 0.7,  # Restricted to copyleft requirements.
-    "lgpl": 0.8,  # Restricted to weak copyleft requirements.
-    "lgpl-2.1": 0.8,  # Restricted to weak copyleft requirements.
-    "lgpl-3.0": 0.8,  # Restricted to weak copyleft requirements.
-    "isc": 0.95,  # Very permissive with minimal restrictions.
-    "h-research": 0.4,  # Restricted to research use only.
-    "intel-research": 0.4,  # Restricted to research use only.
-    "lppl-1.3c": 0.5,  # Restricted to LaTeX-related requirements.
-    "ms-pl": 0.85,  # Permissive but with some restrictions on Microsoft's requirements.
-    "apple-ascl": 0.85,  # Permissive but with some restrictions on Apple's requirements.
-    "apple-amlr": 0.6,  # Restricted by Apple's requirements.
-    "mpl-2.0": 0.8,  # Permissive but with some restrictions on Mozilla's requirements.
-    "odc-by": 0.7,  # Permissive but with some restrictions on data sharing.
-    "odbl": 0.95,  # Very permissive with minimal restrictions.
-    "openmdw-1.0": 0.7,  # Permissive but with some restrictions on data sharing.
-    "openrail++": 0.8,  # Similar to OpenRail but with some additional restrictions.
-    "osl-3.0": 0.85,  # Permissive but with some restrictions on derivative works.
-    "postgresql": 0.85,  # Permissive but with some restrictions on PostgreSQL's requirements.
-    "ofl-1.1": 0.95,  # Very permissive with minimal restrictions.
-    "ncsa": 0.95,  # Very permissive with minimal restrictions.
-    "unlicense": 1.0,  # Public domain dedication with no restrictions.
-    "zlib": 0.95,  # Very permissive with minimal restrictions.
-    "pddl": 0.99,  # Public domain dedication with no restrictions.
-    "lgpl-lr": 0.6,  # Restricted to weak copyleft requirements.
-    "deepfloyd-if-license": 0.7,  # Permissive but with some restrictions on data sharing.
-    "fair-noncommercial-research-license": 0.4,  # Restricted to non-commercial research use.
-    "llama2": 0.7,  # Permissive but with some restrictions on Meta's requirements.
-    "llama3": 0.7,  # Permissive but with some restrictions on Meta's requirements.
-    "llama3.1": 0.7,  # Permissive but with some restrictions on Meta's requirements.
-    "llama3.2": 0.7,  # Permissive but with some restrictions on Meta's requirements.
-    "llama3.3": 0.7,  # Permissive but with some restrictions on Meta's requirements.
-    "llama4": 0.7,  # Permissive but with some restrictions on Meta's requirements.
-    "grok2-community": 0.7,  # Permissive but with some restrictions on data sharing.
-    "gemma": 0.7,  # Permissive but with some restrictions on Google's requirements.
-}
 # Try to import GitPython, fallback to subprocess if not available
 try:
     import git
@@ -261,7 +181,7 @@ def analyze_model_repository(model_name: str, model_url: str, model_type: str = 
                 # else:
                 #     repo_url = f"https://huggingface.co/{model_name}.git"
 
-                print(f"Cloning repository: {repo_url}")
+                print(f"Cloning repository: {model_url}")
 
                 # Clone repository
                 if GIT_PYTHON_AVAILABLE:
@@ -527,15 +447,8 @@ def calculate_metrics(data: Dict[str, Any], category: UrlCategory, code_url: Opt
     likes = data.get('likes', 0)
     has_card = bool(data.get('cardData') or data.get('has_model_card'))
 
-    # License (check tags)
-    tags = data.get('tags', [])
-    start_time = time.perf_counter()
-    license_str = data.get('cardData').get('license') if data.get('cardData') is not None else "unknown"
-    if isinstance(license_str, list):
-        license_str = license_str[0]
-    license_score: float = license_score_map.get(license_str,0) if (license_str is not None and license_str != "unknown" and license_str != "other") else 0.0
-    end_time = time.perf_counter()
-    license_latency = max(10, round((end_time - start_time) * 1000) + 10)  # Add base latency
+    # License calculation with timing
+    license_score, license_latency = calculate_license_score_with_timing(data)
 
     # Ramp-up time - enhanced calculation with timing
     ramp_up, ramp_up_latency = calculate_ramp_up_time_with_timing(data, model_name)
@@ -550,11 +463,7 @@ def calculate_metrics(data: Dict[str, Any], category: UrlCategory, code_url: Opt
     dataset_qual, dataset_qual_latency = calculate_dataset_quality_with_timing(data, downloads, likes)
 
     # Code quality using Flake8 analysis
-    if code_url is not None:
-        code_qual, code_qual_latency = calculate_code_quality_with_flake8(code_url, model_name)
-    else:
-        code_qual = 0
-        code_qual_latency = 0
+    code_qual, code_qual_latency = calculate_code_quality_with_timing(code_url, model_name)
 
     # Net score will be calculated separately with complete metrics including bus_factor and size_score
 
@@ -879,246 +788,6 @@ def score_url(url: str, category: UrlCategory, code_url: Optional[str] = None) -
 #        )
 
 
-def run_flake8_on_repo(repo_path: str) -> tuple[float, int]:
-    """
-    Run Flake8 on a code repository and calculate quality score.
-
-    Args:
-        repo_path: Path to the repository directory
-
-    Returns:
-        tuple of (quality_score, latency_ms)
-    """
-    start_time = time.time()
-
-    try:
-        # First, check if there are any Python files in the repository
-        python_files = []
-        for root, dirs, files in os.walk(repo_path):
-            for file in files:
-                if file.endswith('.py'):
-                    python_files.append(os.path.join(root, file))
-
-        # If no Python files found, return 0.0 (no code to analyze)
-        if not python_files:
-            quality_score = 0.0
-        else:
-            # Run flake8 on the repository
-            result = subprocess.run(
-                ['python3.9', '-m', 'flake8', repo_path, '--count', '--statistics'],
-                capture_output=True,
-                text=True,
-                timeout=30
-            )
-
-            # Parse flake8 output to get error counts
-            output_lines = result.stdout.strip().split('\n')
-            total_errors = 0
-
-            # Extract error count from the last line if it exists
-            if output_lines and output_lines[-1]:
-                try:
-                    # The last line should contain just the total count (e.g., "369")
-                    # If it's a number, that's our total error count
-                    last_line = output_lines[-1].strip()
-                    if last_line.isdigit():
-                        total_errors = int(last_line)
-                    else:
-                        # Fallback: look for pattern like "1     E" or "5     W" etc.
-                        error_match = re.search(r'(\d+)\s+[EW]', last_line)
-                        if error_match:
-                            total_errors = int(error_match.group(1))
-                except (ValueError, IndexError):
-                    pass
-
-            # Calculate quality score based on error count
-            # More lenient scoring that better reflects real-world code quality
-            if total_errors == 0:
-                quality_score = 1.0  # Perfect code (0 errors)
-            elif total_errors <= 10:
-                quality_score = 0.9  # Excellent (1-10 errors)
-            elif total_errors <= 50:
-                quality_score = 0.8  # Very good (11-50 errors)
-            elif total_errors <= 100:
-                quality_score = 0.7  # Good (51-100 errors)
-            elif total_errors <= 200:
-                quality_score = 0.6  # Fair (101-200 errors)
-            elif total_errors <= 500:
-                quality_score = 0.5  # Below average (201-500 errors)
-            elif total_errors <= 1000:
-                quality_score = 0.4  # Poor (501-1000 errors)
-            elif total_errors <= 2000:
-                quality_score = 0.3  # Very poor (1001-2000 errors)
-            else:
-                quality_score = 0.2  # Extremely poor (2000+ errors)
-
-    except subprocess.TimeoutExpired:
-        quality_score = 0.0
-    except Exception:
-        quality_score = 0.0
-
-    end_time = time.time()
-    latency_ms = int((end_time - start_time) * 1000)
-
-    return quality_score, latency_ms
-
-
-def find_code_repo_via_genai(model_name: str) -> Optional[str]:
-    """
-    Use Purdue GenAI Studio API to find code repository link in model README.
-
-    Uses the official Purdue GenAI Studio API endpoint:
-    https://genai.rcac.purdue.edu/api/chat/completions
-
-    Args:
-        model_name: Name of the model (e.g., "google-bert/bert-base-uncased")
-
-    Returns:
-        Code repository URL if found, None otherwise
-    """
-    try:
-        # Purdue GenAI Studio API endpoint (from official documentation)
-        api_url = "https://genai.rcac.purdue.edu/api/chat/completions"
-
-        # Prepare the prompt to find code repository in model README
-        prompt = f"""
-        Please analyze the Hugging Face model "{model_name}" and find any code repository links in its README or documentation.
-
-        Look for:
-        - GitHub repository links
-        - GitLab repository links
-        - Other code hosting platform links
-        - Source code references
-
-        Return only the URL if found, or "NO_CODE_FOUND" if no code repository is found.
-        """
-
-        headers = {
-            "Authorization": "Bearer sk-ed2de44f587645c5b3fe62bb8f2328fc",
-            "Content-Type": "application/json"
-        }
-
-        payload = {
-            "model": "llama3.1:latest",  # Using the model specified in Purdue GenAI Studio docs
-            "messages": [
-                {
-                    "role": "user",
-                    "content": prompt
-                }
-            ],
-            "stream": False,  # Non-streaming response as per documentation
-            "max_tokens": 200,
-            "temperature": 0.1
-        }
-
-        response = requests.post(api_url, headers=headers, json=payload, timeout=30)
-
-        if response.status_code == 200:
-            result = response.json()
-            content = result.get('choices', [{}])[0].get('message', {}).get('content', '').strip()
-
-            # Check if a valid URL was returned
-            if content and content != "NO_CODE_FOUND":
-                # Extract URLs from the response text
-                import re
-                url_pattern = r'https://(?:github\.com|gitlab\.com|bitbucket\.org|sourceforge\.net)/[^\s\)]+'
-                urls = re.findall(url_pattern, content)
-
-                if urls:
-                    # Return the first valid URL found
-                    return urls[0]
-        else:
-            loggerInstance.logger.log_info(f"Purdue GenAI Studio API returned status {response.status_code}: {response.text}")
-
-        # Fallback: try to extract GitHub repo from model name
-        if '/' in model_name:
-            owner, repo = model_name.split('/', 1)
-            github_url = f"https://github.com/{owner}/{repo}"
-
-            # Test if the URL exists
-            test_response = requests.get(github_url, timeout=5)
-            if test_response.status_code == 200:
-                return github_url
-
-        return None
-
-    except Exception as e:
-        # Log the error but don't fail completely
-        loggerInstance.logger.log_info(f"GenAI API error: {e}")
-
-        # Fallback: try to extract GitHub repo from model name
-        try:
-            if '/' in model_name:
-                owner, repo = model_name.split('/', 1)
-                github_url = f"https://github.com/{owner}/{repo}"
-
-                test_response = requests.get(github_url, timeout=5)
-                if test_response.status_code == 200:
-                    return github_url
-        except Exception:
-            pass
-
-        return None
-
-
-def calculate_code_quality_with_flake8(code_url: str, model_name: str) -> tuple[float, int]:
-    """
-    Calculate code quality using Flake8, with fallback to GenAI API.
-
-    Args:
-        code_url: Direct code URL if available
-        model_name: Model name for fallback search
-
-    Returns:
-        tuple of (quality_score, latency_ms)
-    """
-    start_time = time.time()
-
-    try:
-        # Clone the repository to a temporary directory
-        with tempfile.TemporaryDirectory() as temp_dir:
-            repo_path = os.path.join(temp_dir, "repo")
-
-            # Clone the repository
-            subprocess.run(
-                ['git', 'clone', '--depth', '1', code_url, repo_path],
-                capture_output=True,
-                timeout=60
-            )
-
-            # Run flake8 on the cloned repository
-            quality_score, _ = run_flake8_on_repo(repo_path)
-
-    except Exception:
-        quality_score = 0.0
-    #else:
-    #    # Fallback: try to find code repo via GenAI API
-    #    code_repo_url = find_code_repo_via_genai(model_name)
-
-    #    if code_repo_url:
-    #        try:
-    #            # Clone and analyze the found repository
-    #            with tempfile.TemporaryDirectory() as temp_dir:
-    #                repo_path = os.path.join(temp_dir, "repo")
-
-    #                subprocess.run(
-    #                    ['git', 'clone', '--depth', '1', code_repo_url, repo_path],
-    #                    capture_output=True,
-    #                    timeout=60
-    #                )
-
-    #                quality_score, _ = run_flake8_on_repo(repo_path)
-
-    #        except Exception:
-    #            quality_score = 0.0
-    #    else:
-    #        # No code repository found, return 0
-    #        quality_score = 0.0
-
-    end_time = time.time()
-    latency_ms = int((end_time - start_time) * 1000)
-
-    return quality_score, latency_ms
 
 
 # Parallel metric computation functions
@@ -1155,18 +824,8 @@ def compute_performance_claims_parallel(data: Dict[str, Any], model_name: str = 
 def compute_license_parallel(data: Dict[str, Any]) -> tuple[str, float, int]:
     """Compute license metric in parallel."""
     try:
-        start_time = time.perf_counter()
-
-        # License calculation
-        card_data = data.get('cardData', {})
-        license_str = card_data.get('license') if card_data else "unknown"
-        if isinstance(license_str, list):
-            license_str = license_str[0]
-        license_score = license_score_map.get(license_str, 0.0) if (license_str is not None and license_str != "unknown" and license_str != "other") else 0.0
-
-        end_time = time.perf_counter()
-        latency = max(10, int((end_time - start_time) * 1000) + 10)
-
+        # Use the license module for calculation
+        license_score, latency = calculate_license_score_with_timing(data)
         return "license", license_score, latency
     except Exception as e:
         loggerInstance.logger.log_info(f"Error computing license: {e}")
@@ -1214,7 +873,7 @@ def compute_dataset_quality_parallel(data: Dict[str, Any], downloads: int, likes
 def compute_code_quality_parallel(code_url: Optional[str], model_name: str) -> tuple[str, float, int]:
     """Compute code_quality metric in parallel."""
     try:
-        score, latency = calculate_code_quality_with_flake8(code_url, model_name)
+        score, latency = calculate_code_quality_with_timing(code_url, model_name)
         return "code_quality", score, latency
     except Exception as e:
         loggerInstance.logger.log_info(f"Error computing code_quality: {e}")
